+++ packages/aws/README.rst
@@ -1,0 +1,7 @@
+rpaframework-aws
+================
+
+This library enables Amazon Web Services (AWS) for `RPA Framework`_
+libraries, such as Textract.
+
+.. _RPA Framework: https://rpaframework.org
