+++ packages/aws/tests/python/test_library.py
@@ -1,0 +1,6 @@
+from RPA.Cloud.AWS import AWS
+
+
+def test_init():
+    lib = AWS()
+    assert lib
