+++ packages/main/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "rpaframework"
-version = "13.0.0"
+version = "13.0.1"
 description = "A collection of tools and libraries for RPA"
 authors = [
 	"RPA Framework <rpafw@robocorp.com>",
@@ -38,6 +38,7 @@ packages = [
 python = "^3.6.2"
 docutils = "*"
 dataclasses = { version = "^0.7", python=">=3.6,<3.7" }
+rpaframework-aws = { version = "^1.0.0", optional = true }
 rpaframework-core = "^7.0.1"
 rpaframework-dialogs = "^1.0.0"
 rpaframework-pdf = "^2.0.0"
@@ -76,8 +77,6 @@ chardet = "^3.0.0"
 PySocks = ">=1.5.6,!=1.5.7,<2.0.0"
 selenium = "^3.141.0"
 click = "^7.1.2"
-boto3 = { version = "^1.13.4", optional = true }
-amazon-textract-response-parser = { version = "^0.1.1", optional = true }
 java-access-bridge-wrapper = "^0.9.4"
 PyYAML = "^5.4.1"
 tenacity = "^8.0.1"
@@ -108,9 +107,9 @@ rpa-crypto = 'RPA.scripts.crypto:main'
 use-robocorp-vault = 'RPA.scripts.robocorp_cloud:main'
 
 [tool.poetry.extras]
+aws = ["rpaframework-aws"]
 cv = ["rpaframework-recognition"]
 playwright = ["robotframework-browser"]
-aws = ["boto3", "amazon-textract-response-parser"]
 
 [tool.black]
 target-version = ["py36", "py37", "py38", "py39"]
