--- packages/main/src/RPA/Cloud/objects/__init__.py
@@ -1,1 +1,0 @@
-from .textract import TextractDocument
